Split up commands tests into files
